ir+metainterp: move IntBound to majit-ir; make_guards stays as extension trait
`IntBound` (data + 80 leaf methods + helpers + Default/Display impls +
the 133-test cfg(test) block) now lives in `majit-ir::intbound`. The one
optimizer-coupled method, `make_guards`, stays in
`metainterp::optimizeopt::intutils` as the `IntBoundMakeGuards`
extension trait so callers continue invoking `bound.make_guards(...)`
after importing the trait. Re-exports keep `crate::optimizeopt::intutils::IntBound`
working at every prior caller; the 5 sites that call `make_guards` also
import the trait.
