+++ majit/majit-ir/src/lib.rs
@@ -3,6 +3,7 @@ pub mod debug;
 pub mod descr;
 pub mod descr_registry;
 pub mod effectinfo;
+pub mod intbound;
 pub mod op_descr;
 pub mod op_type_index;
 pub mod optimize;
