+++ majit/majit-metainterp/src/optimizeopt/virtualstate.rs
@@ -145,7 +145,7 @@ pub(crate) fn info_type_matches(expected: Type, incoming: &VirtualStateInfo) -> 
 
 use crate::optimizeopt::OptContext;
 use crate::optimizeopt::info::PtrInfo;
-use crate::optimizeopt::intutils::IntBound;
+use crate::optimizeopt::intutils::{IntBound, IntBoundMakeGuards};
 
 /// Abstract info for one value at the loop boundary.
 ///
